+++ nano/node/bootstrap/account_sets.hpp
@@ -2,7 +2,8 @@
 
 #include <nano/lib/numbers.hpp>
 #include <nano/node/bootstrap/bootstrap_config.hpp>
-#include <nano/node/bootstrap_ascending/common.hpp>
+#include <nano/node/bootstrap/common.hpp>
+#include <nano/node/fwd.hpp>
 
 #include <boost/multi_index/hashed_index.hpp>
 #include <boost/multi_index/mem_fun.hpp>
