+++ nano/node/bootstrap/common.hpp
