+++ nano/node/bootstrap/service.cpp
@@ -3,8 +3,8 @@
 #include <nano/lib/stats_enums.hpp>
 #include <nano/lib/thread_roles.hpp>
 #include <nano/node/blockprocessor.hpp>
-#include <nano/node/bootstrap_ascending/crawlers.hpp>
-#include <nano/node/bootstrap_ascending/service.hpp>
+#include <nano/node/bootstrap/crawlers.hpp>
+#include <nano/node/bootstrap/service.hpp>
 #include <nano/node/network.hpp>
 #include <nano/node/nodeconfig.hpp>
 #include <nano/node/transport/transport.hpp>
