+++ nano/node/message_processor.cpp
@@ -1,5 +1,5 @@
 #include <nano/lib/thread_roles.hpp>
-#include <nano/node/bootstrap_ascending/service.hpp>
+#include <nano/node/bootstrap/service.hpp>
 #include <nano/node/message_processor.hpp>
 #include <nano/node/node.hpp>
 #include <nano/node/telemetry.hpp>
